+++ backend/internal/service/handler/reviews/handler.go
@@ -1,0 +1,15 @@
+package reviews
+
+import "platnm/internal/storage"
+
+type ReviewHandler struct {
+	reviewRepository storage.ReviewRepository
+	userRepository storage.UserRepository
+}
+
+func NewReviewHandler(reviewRepository storage.ReviewRepository, userRepository storage.UserRepository) *ReviewHandler {
+	return &ReviewHandler{
+		reviewRepository,
+		userRepository,
+	}
+}
