+++ backend/internal/service/handler/reviews/review.go
@@ -1,24 +1,10 @@
 package reviews
 
 import (
-	"platnm/internal/storage"
-
 	"github.com/gofiber/fiber/v2"
 
 	"platnm/internal/errs"
 )
-
-type ReviewHandler struct {
-	reviewRepository storage.ReviewRepository
-	userRepository storage.UserRepository
-}
-
-func NewReviewHandler(reviewRepository storage.ReviewRepository, userRepository storage.UserRepository) *ReviewHandler {
-	return &ReviewHandler{
-		reviewRepository,
-		userRepository,
-	}
-}
 
 func (h *ReviewHandler) GetReviewsByUserID(c *fiber.Ctx) error {
 
@@ -27,14 +13,14 @@ func (h *ReviewHandler) GetReviewsByUserID(c *fiber.Ctx) error {
 	exists, err := h.reviewRepository.UserExists(c.Context(), id)
 
 	if err != nil {
-		return err 
+		return err
 	}
 
 	if !exists {
 		return errs.NotFound("User", "id", id)
 	}
 
-	review,err := h.reviewRepository.GetReviewsByUserID(c.Context(), id)
+	review, err := h.reviewRepository.GetReviewsByUserID(c.Context(), id)
 
 	if err != nil {
 		return err
